Create README.md
Deploying a script to process PCAP files on arrival using Snort IDS.

The following script will check to see if it is already running by using a lockfile function. If it is running it will exit to avoid overlapping processing. If it is not already running it checks to see if there are any files which have arrived in /home/pcap/filedrop. If any files have arrived, it creates a uniquely named temporary working directory. It then checks to see if the file is open by another program in case it is a large file still being written to. If the file is not still being written to, it moves it from the secure dropoff point into the temporary working directory it has created and processes it. Finally it cleans up after itself and releases the lockfile when it is finished, thereby enabling it to rerun again without overlap.
https://www.coresentinel.com/processing-pcap-files-snort/